# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Smart Retry Orchestrator API
  description: |
    The Smart Retry Orchestrator API provides intelligent retry logic for failed
    payment transactions. It classifies failures, determines optimal retry
    strategies, tracks processor health, and maximizes revenue recovery while
    minimizing unnecessary retry costs.

    ## Core Concepts

    - **Failure Classification**: Each incoming failure event is classified as either
      a HARD_DECLINE (non-retryable) or SOFT_DECLINE (retryable).
    - **Retry Strategy**: For soft declines the orchestrator selects a strategy
      (IMMEDIATE, DELAYED, or ALTERNATIVE_PROCESSOR) based on adaptive weights
      and processor health.
    - **Processor Health**: Processor success rates are continuously monitored.
      Processors transition between HEALTHY, DEGRADED, and DOWN states.
    - **Recovery Metrics**: Aggregated statistics on recovery rates, revenue at
      risk, revenue recovered, and retry costs.
  version: "1.0.0"
  contact:
    name: Smart Retry Orchestrator Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /docs/openapi.yaml:
    get:
      operationId: serveOpenAPI
      summary: OpenAPI specification
      description: Returns the raw OpenAPI YAML specification file.
      tags:
        - System
      responses:
        "200":
          description: OpenAPI spec returned.
          content:
            application/yaml:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Classification
  # ---------------------------------------------------------------------------
  /api/v1/failures:
    post:
      operationId: submitFailure
      summary: Submit a single failure event
      description: |
        Accepts a single payment failure event, classifies it, determines whether
        a retry should be attempted, executes the retry chain, and returns the
        full result including classification, retry attempts, and final status.
      tags:
        - Classification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FailureRequest"
            example:
              transaction_id: "txn_abc123"
              amount: 99.99
              currency: "USD"
              failure_code: "NETWORK_ERROR"
              processor: "PROCESSOR_A"
              timestamp: "2026-02-20T10:30:00Z"
      responses:
        "201":
          description: Failure event classified and processed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessFailureResult"
              example:
                transaction_id: "txn_abc123"
                classification:
                  decline_type: SOFT_DECLINE
                  strategy: IMMEDIATE
                  failure_code: NETWORK_ERROR
                  reasoning: "Network error is transient; immediate retry recommended."
                should_retry: true
                retry_attempts:
                  - attempt_number: 1
                    processor: "PROCESSOR_A"
                    strategy: IMMEDIATE
                    success: true
                    cost: 0.30
                    delay_seconds: 0
                    reasoning: "Immediate retry on same processor PROCESSOR_A (state: HEALTHY, cost: $0.30)"
                    timestamp: "2026-02-20T10:30:01Z"
                final_status: RECOVERED
                total_cost: 0.30
                reasoning: "Recovered after 1 attempt(s). Final successful attempt on PROCESSOR_A. Total cost: $0.30"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/failures/batch:
    post:
      operationId: submitFailureBatch
      summary: Submit a batch of failure events
      description: |
        Accepts a JSON object with an "events" array of payment failure events,
        classifies each one, and returns aggregated results. Partial failures
        are reported per-event inside the results array; the overall HTTP status
        is 201 as long as the batch was accepted.
      tags:
        - Classification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchFailureRequest"
            example:
              events:
                - transaction_id: "txn_batch_001"
                  amount: 50.00
                  currency: "USD"
                  failure_code: "NETWORK_ERROR"
                  processor: "PROCESSOR_A"
                - transaction_id: "txn_batch_002"
                  amount: 120.00
                  currency: "EUR"
                  failure_code: "CARD_EXPIRED"
                  processor: "PROCESSOR_B"
                - transaction_id: "txn_batch_003"
                  amount: 75.50
                  currency: "USD"
                  failure_code: "ISSUER_UNAVAILABLE"
                  processor: "PROCESSOR_C"
      responses:
        "201":
          description: Batch processed. Individual errors are reported per-item in the results array.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResponse"
              example:
                total: 3
                processed: 2
                errors: 1
                results:
                  - transaction_id: "txn_batch_001"
                    status: "RECOVERED"
                    should_retry: true
                    error: ""
                  - transaction_id: "txn_batch_002"
                    status: "HARD_DECLINED"
                    should_retry: false
                    error: ""
                  - transaction_id: "txn_batch_003"
                    status: "error"
                    should_retry: false
                    error: "transaction ID already exists"
                summary:
                  hard_declines: 1
                  soft_declines: 1
                  recovered: 1
                  exhausted: 0
                  retrying: 0
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Retry Outcome
  # ---------------------------------------------------------------------------
  /api/v1/retries/{transactionID}/outcome:
    post:
      operationId: reportRetryOutcome
      summary: Report a retry outcome
      description: |
        Reports the outcome of a retry attempt for the given transaction. The
        orchestrator updates internal state, processor health, adaptive weights,
        and determines whether the transaction was recovered.
      tags:
        - Transactions
      parameters:
        - $ref: "#/components/parameters/TransactionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryOutcomeRequest"
            example:
              processor: "PROCESSOR_B"
              success: true
      responses:
        "200":
          description: Retry outcome recorded successfully.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    description: Confirmation that the outcome was recorded.
                    example: recorded
              example:
                status: recorded
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Transactions
  # ---------------------------------------------------------------------------
  /api/v1/transactions:
    get:
      operationId: listTransactions
      summary: List transactions
      description: |
        Returns a paginated list of transactions, optionally filtered by status
        and/or processor.
      tags:
        - Transactions
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by transaction status.
          schema:
            $ref: "#/components/schemas/TransactionStatus"
        - name: processor
          in: query
          required: false
          description: Filter by original processor.
          schema:
            $ref: "#/components/schemas/ProcessorName"
        - name: limit
          in: query
          required: false
          description: Maximum number of transactions to return.
          schema:
            type: integer
            minimum: 1
            default: 50
        - name: offset
          in: query
          required: false
          description: Number of transactions to skip for pagination.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of transactions returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/transactions/{transactionID}:
    get:
      operationId: getTransaction
      summary: Get full transaction details
      description: |
        Returns the complete transaction record including the original failure
        event, classification, current status, retry attempts, and cost totals.
      tags:
        - Transactions
      parameters:
        - $ref: "#/components/parameters/TransactionID"
      responses:
        "200":
          description: Transaction found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
              example:
                id: "txn_abc123"
                original_event:
                  transaction_id: "txn_abc123"
                  amount: 99.99
                  currency: "USD"
                  failure_code: "NETWORK_ERROR"
                  processor: "PROCESSOR_A"
                  timestamp: "2026-02-20T10:30:00Z"
                classification:
                  decline_type: SOFT_DECLINE
                  strategy: IMMEDIATE
                  failure_code: NETWORK_ERROR
                  reasoning: "Network error is transient; immediate retry recommended."
                status: RECOVERED
                retry_count: 1
                max_retries: 3
                retry_attempts:
                  - attempt_number: 1
                    processor: "PROCESSOR_B"
                    strategy: ALTERNATIVE_PROCESSOR
                    success: true
                    cost: 0.25
                    delay_seconds: 0
                    reasoning: "Routed to alternative processor due to degraded primary."
                    timestamp: "2026-02-20T10:31:00Z"
                total_cost: 0.25
                created_at: "2026-02-20T10:30:00Z"
                updated_at: "2026-02-20T10:31:00Z"
                recovered_at: "2026-02-20T10:31:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/transactions/{transactionID}/retries:
    get:
      operationId: getRetryHistory
      summary: Get retry history for a transaction
      description: |
        Returns the retry history for a transaction, including the transaction ID,
        retry count, all retry attempts, and all decision logs.
      tags:
        - Transactions
      parameters:
        - $ref: "#/components/parameters/TransactionID"
      responses:
        "200":
          description: Retry history returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetryHistoryResponse"
              example:
                transaction_id: "txn_abc123"
                retry_count: 2
                attempts:
                  - attempt_number: 1
                    processor: "PROCESSOR_A"
                    strategy: IMMEDIATE
                    success: false
                    cost: 0.30
                    delay_seconds: 0
                    reasoning: "Immediate retry on same processor."
                    timestamp: "2026-02-20T10:31:00Z"
                  - attempt_number: 2
                    processor: "PROCESSOR_B"
                    strategy: ALTERNATIVE_PROCESSOR
                    success: true
                    cost: 0.25
                    delay_seconds: 0
                    reasoning: "Routed to alternative processor after first retry failed."
                    timestamp: "2026-02-20T10:32:00Z"
                decisions:
                  - transaction_id: "txn_abc123"
                    attempt_number: 1
                    source_processor: "PROCESSOR_A"
                    target_processor: "PROCESSOR_A"
                    strategy: IMMEDIATE
                    reasoning: "Immediate retry on same processor."
                    delay_seconds: 0
                    timestamp: "2026-02-20T10:31:00Z"
                  - transaction_id: "txn_abc123"
                    attempt_number: 2
                    source_processor: "PROCESSOR_A"
                    target_processor: "PROCESSOR_B"
                    strategy: ALTERNATIVE_PROCESSOR
                    reasoning: "Routed to alternative processor."
                    delay_seconds: 0
                    timestamp: "2026-02-20T10:32:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Processors
  # ---------------------------------------------------------------------------
...
  /api/v1/strategy:
    get:
      operationId: getAdaptiveStrategy
      summary: Get adaptive strategy weights
      description: |
        Returns the current adaptive strategy weights used by the orchestrator.
        These are per-error-code, per-processor success rates learned from
        retry outcomes. The response contains an array of adaptive weights
        and a description.
      tags:
        - Strategy
      responses:
        "200":
          description: Current adaptive strategy weights returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveStrategyResponse"
              example:
                adaptive_weights:
                  - failure_code: "NETWORK_ERROR"
                    weights:
                      PROCESSOR_A: 0.85
                      PROCESSOR_B: 0.70
                      PROCESSOR_C: 0.65
                    sample_count:
                      PROCESSOR_A: 50
                      PROCESSOR_B: 30
                      PROCESSOR_C: 20
                  - failure_code: "PROCESSOR_TIMEOUT"
                    weights:
                      PROCESSOR_A: 0.80
                      PROCESSOR_B: 0.75
                    sample_count:
                      PROCESSOR_A: 40
                      PROCESSOR_B: 25
                description: "Per-error-code, per-processor success rates learned from retry outcomes"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Admin / Testing
  # ---------------------------------------------------------------------------
  /api/v1/test/generate:
    post:
      operationId: generateTestEvents
      summary: Generate test failure events
      description: |
        Generates 210 deterministic synthetic failure events with randomized
        amounts, currencies, failure codes, and processors. Processes all events
        through the orchestrator. Useful for populating the system during
        development and testing.
      tags:
        - Admin
      responses:
        "201":
          description: Test events generated and processed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestGenerateResponse"
              example:
                status: generated
                count: 210
                message: "Test events generated and processed successfully"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/test/reset:
    post:
      operationId: resetTestData
      summary: Clear all data
      description: |
        Removes all transactions, retry history, processor health state, and
        metrics from the system. Intended for use in test and development
        environments only.
      tags:
        - Admin
      responses:
        "200":
          description: All data cleared successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResetResponse"
              example:
                status: reset
                message: "All data cleared successfully"
        "500":
          $ref: "#/components/responses/InternalServerError"

# =============================================================================
# Components
# =============================================================================
components:
...
  responses:
    BadRequest:
      description: The request body or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "VALIDATION_ERROR"
            messages:
              - "transaction_id is required"
              - "amount must be greater than 0"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "NOT_FOUND"
            messages:
              - "transaction not found"
    Conflict:
      description: A resource with the same identifier already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "DUPLICATE_TRANSACTION"
            messages:
              - "transaction ID already exists"
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "PROCESSING_ERROR"
            messages:
              - "an unexpected error occurred"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    TransactionStatus:
      type: string
      description: Current lifecycle status of a transaction.
      enum:
        - PENDING
        - HARD_DECLINED
        - RETRYING
        - RECOVERED
        - EXHAUSTED
      example: RETRYING
...
    FailureRequest:
      type: object
      description: A payment failure event submitted for classification and retry.
      required:
        - transaction_id
        - amount
        - currency
        - failure_code
        - processor
      properties:
        transaction_id:
          type: string
          description: Unique identifier for the transaction.
          example: "txn_abc123"
        amount:
          type: number
          format: double
          description: Transaction amount. Must be greater than 0.
          minimum: 0.01
          example: 99.99
        currency:
          type: string
          description: "ISO 4217 currency code. Accepted values: IDR, MYR, PHP, USD, EUR, GBP, SGD."
          example: "USD"
        failure_code:
          type: string
          description: The failure reason code from the processor.
          example: "NETWORK_ERROR"
        processor:
          type: string
          description: The processor that reported the failure.
          example: "PROCESSOR_A"
        timestamp:
          type: string
          format: date-time
          description: "Optional ISO 8601 / RFC3339 timestamp. Defaults to current time if omitted."
          example: "2026-02-20T10:30:00Z"
    BatchFailureRequest:
      type: object
      description: Wrapper object containing an array of failure events for batch processing.
      required:
        - events
      properties:
        events:
          type: array
          description: Array of failure events to process.
          minItems: 1
          items:
            $ref: "#/components/schemas/FailureRequest"
...
    ProcessFailureResult:
      type: object
      description: The full result of processing a failure event through the orchestrator.
      required:
        - transaction_id
        - classification
        - should_retry
        - final_status
        - total_cost
        - reasoning
      properties:
        transaction_id:
          type: string
          description: Transaction identifier.
          example: "txn_abc123"
        classification:
          $ref: "#/components/schemas/ClassificationResult"
        should_retry:
          type: boolean
          description: Whether a retry was or should be attempted.
          example: true
        retry_attempts:
          type: array
          description: Ordered list of retry attempts executed (empty for hard declines).
          items:
            $ref: "#/components/schemas/RetryAttempt"
        final_status:
          $ref: "#/components/schemas/TransactionStatus"
        total_cost:
          type: number
          format: double
          description: Cumulative cost of all retry attempts for this transaction.
          minimum: 0
          example: 0.30
        reasoning:
          type: string
          description: Human-readable explanation of the processing decision.
          example: "Recovered after 1 attempt(s). Final successful attempt on PROCESSOR_A. Total cost: $0.30"
...
    BatchResponse:
      type: object
      description: Response from submitting a batch of failure events.
      required:
        - total
        - processed
        - errors
        - results
        - summary
      properties:
        total:
          type: integer
          description: Total number of events in the submitted batch.
          minimum: 0
          example: 3
        processed:
          type: integer
          description: Number of events that were successfully processed.
          minimum: 0
          example: 2
        errors:
          type: integer
          description: Number of events that encountered errors.
          minimum: 0
          example: 1
        results:
          type: array
          description: Per-event results with status and optional error.
          items:
            $ref: "#/components/schemas/BatchItemResult"
        summary:
          $ref: "#/components/schemas/BatchSummary"
...